Extend forecast
I made changes to various pieces of code to allow the forecast to go 72
hours instead of 24.  Backfill runs should still go only 24.  One
inefficiency is that I now always get 72 hours of atm forcing.  To fix
this I would have to pass the run_type to the mat worker, which is
inconsistent with the current architecture.