run type info to worker
I recoded Lfun.run_worker() to pass “run_type” to the workers, and then
recoded all make_forcing_worker.m instances to accept this.  The
motivation was so that for backfill I would just do 24 hours of atm
forcing.  This also helps when using the 20140107 synthetic files, but
it will help for most of the older archive too.