HYCOM bug fix
Mainly this was adding a bit of (crappy) code to Ofun.py in the
function process_extra.  The goal was to patch over some bad salinity
fields.

Caution: this may crash when run in forecast mode if the bad day is at
the start or end time of an extraction.