driver_roms1
Edited driver_roms1.sh in the same way that I did for
driver_forcing1.sh, so that it can be run again (e.g. by cron) but only
run ROMS if needed (can override with -c).

The point of doing this is so that I can have cron set to try several
times, as part of my ongoing reliability campaign.